refact: LyricsContainer 컴포넌트 분리
