+++ components/study/LyricsContainer/LyricsContent/TextSizeController.tsx
