+++ components/study/LyricsContainer/Steps.tsx
